test_identities: Use starfish instead of 3-to-1 ellipse.
This permits us to use a lower FMM order. My guess is QBX finds the
starfish more difficult than the ellipse, so that in the starfish the
QBX error dominates compared to FMM error for the values of h we are
using.

This change should also make the build green.

:fireworks: