Introduce a smarter quadrature resolution measure
This measure is aware of element aspect-ratios in 3D and and varying
'parametrization speeds' across an element. It's also computed using
the regular expression evaluator, not with custom machinery. The
evaluator was made smarter as needed--such as by allowing it to operate
on the 'stage 2' discretization of a layer potential source.

This quadrature resolution measure replaces panel sizes where appropriate.